fix iOS 12.4 triangulared view disabled state
